+++ osgHelper/include/osgHelper/ppu/BlendTexture.h
@@ -29,8 +29,8 @@ namespace osgHelper
 
       RenderTextureUnitSink getBlendTextureSink();
 
-      void setResolution(const osg::Vec2f& resolution);
-      void onResizeViewport(const osg::Vec2f& resolution) override;
+      void setResolution(const osg::Vec2i& resolution);
+      void onResizeViewport(const osg::Vec2i& resolution) override;
 
     protected:
       Status initializeUnits(const osg::GL2Extensions* extensions) override;
